+++ backend/src/bcws/dto/update-bcws-personnel.dto.ts
@@ -1,6 +1,5 @@
-import { PartialType } from '@nestjs/swagger';
+import { PartialType } from '@nestjs/mapped-types';
 import { CreatePersonnelBcwsDTO } from './create-bcws-personnel.dto';
-
 export class UpdateBcwsPersonnelDTO extends PartialType(
   CreatePersonnelBcwsDTO,
 ) {}
