+++ backend/src/personnel/dto/details/update-personnel-details.dto.ts
@@ -1,0 +1,6 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { PersonnelDetailsDTO } from './personnel-details.dto';
+
+export class UpdatePersonnelDetailsDTO extends PartialType(
+  PersonnelDetailsDTO,
+) {}
