+++ backend/src/personnel/dto/emergency-contact/update-emergency-contact.ts
@@ -1,0 +1,6 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { EmergencyContactInfo } from './emergency-contact.dto';
+
+export class UpdateEmergencyContactInfo extends PartialType(
+  EmergencyContactInfo,
+) {}
