+++ backend/src/personnel/dto/employment-info/update-employment-info.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { EmploymentInfo } from './employment-info.dto';
+
+export class UpdateEmploymentInfo extends PartialType(EmploymentInfo) {}
