+++ backend/src/personnel/dto/index.ts
@@ -1,5 +1,5 @@
 export * from './create-personnel.dto';
-export * from './get-availability.dto';
+export * from './availability/get-availability.dto';
 export * from './get-personnel.dto';
-export * from './update-availability.dto';
+export * from './availability/update-availability.dto';
 export * from './update-personnel.dto';
