+++ backend/src/personnel/dto/skills/create-personnel-languages.dto.ts
@@ -3,7 +3,7 @@ import { IsEnum } from 'class-validator';
 import {
   LanguageLevelType,
   LanguageProficiency,
-} from '../../common/enums/bcws';
+} from '../../../common/enums/bcws';
 
 export class CreatePersonnelLanguagesDTO {
   @ApiProperty({
