+++ backend/src/personnel/dto/skills/update-personnel-certifications.dto.ts
@@ -1,0 +1,6 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateCertificationsDTO } from './create-certifications.dto';
+
+export class UpdateCertificationsDTO extends PartialType(
+  CreateCertificationsDTO,
+) {}
