+++ backend/src/personnel/dto/skills/update-personnel-skills.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { SkillsDTO } from './create-personnel-skills-dto';
+
+export class UpdateSkillsDTO extends PartialType(SkillsDTO) {}
