+++ backend/src/personnel/dto/skills/update-personnel-tools.dto.ts
@@ -1,0 +1,10 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { CreatePersonnelToolsDTO } from './create-personnel-tools.dto';
+
+export class UpdatePersonnelToolsDTO extends CreatePersonnelToolsDTO {
+  @ApiProperty({
+    description: 'Tool id',
+    required: true,
+  })
+  toolId: number;
+}
