+++ backend/src/personnel/dto/supervisor/update-supervisor.dto.ts
@@ -1,0 +1,6 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { SupervisorInformationDTO } from './create-supervisor-info.dto';
+
+export class UpdateSupervisorInformationDTO extends PartialType(
+  SupervisorInformationDTO,
+) {}
