+++ backend/src/personnel/update-preferences.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/swagger';
+import { UpdatePersonnelDTO } from './dto/update-personnel.dto';
+
+export class UpdatePreferencesDTO extends PartialType(UpdatePersonnelDTO) {}
